Добавлена аутентификация по токену

+++ app/controllers/api/v1/channels_controller.rb
@@ -1,4 +1,5 @@
 class Api::V1::ChannelsController < Api::V1::BaseController
+  before_action :authenticate_with_token!
 
   def index
     @channels = Channel.order(id: :DESC).page(params[:page])
@@ -17,13 +18,22 @@ class Api::V1::ChannelsController < Api::V1::BaseController
   end
 
    def destroy
-    channel = Channel.find(params[:id])
 
-    if !channel.destroy
-      return render text: "500 Internal Server Error", status: 500
-    end
+    channel = Channel.where(id: params[:id]).first
 
-    head status: 204
+    if channel.user == current_user
+      if channel
+        if !channel.destroy
+          head status: 204
+        else
+          render text: "500 Internal Server Error", status: 500
+        end
+      else  
+        render text: "404 Not Found", status: 404
+      end
+    else
+      render text: "403 Forbidden ", status: 403
+    end  
   end
 
   private
